feat: confirmation before remove

+++ src/app/team-stats/team-stats.component.ts
@@ -3,6 +3,7 @@ import { Observable, tap } from 'rxjs';
 
 import { Game, Stats, Team } from '../data.models';
 import { NbaService } from '../nba.service';
+import { ModalService } from '../widges/modal';
 
 @Component({
   selector: 'app-team-stats',
@@ -15,7 +16,8 @@ export class TeamStatsComponent implements OnInit {
 
   games$!: Observable<Game[]>;
   stats!: Stats;
-  constructor(protected nbaService: NbaService) {}
+
+  constructor(protected nbaService: NbaService, private modal: ModalService) {}
 
   ngOnInit(): void {
     this.games$ = this.nbaService
@@ -28,7 +30,14 @@ export class TeamStatsComponent implements OnInit {
       );
   }
 
-  removeTrackedTeam(team: Team) {
-    this.nbaService.removeTrackedTeam(team);
+  async removeTrackedTeam(team: Team) {
+    const confirm = await this.modal.show(
+      'Untrack Team',
+      `Are you sure want to remove team "${team.full_name}"?`,
+      'YesNo'
+    );
+    if (confirm == 'Yes') {
+      this.nbaService.removeTrackedTeam(team);
+    }
   }
 }
